requirements/base.txt
Django==3.2
django-environ==0.4.5
pytz==2020.5
sqlparse==0.4.1
Pillow==8.1.0
djangorestframework==3.12.4
psycopg2-binary==2.8.6
django-rest-swagger==2.2.0
drf-yasg==1.20.0
